+++ orflib-0.10.0/pyorflib/pyfunctions4.hpp
@@ -7,7 +7,6 @@
 #include <orflib/market/market.hpp>
 #include <orflib/products/europeancallput.hpp>
 #include <orflib/products/americancallput.hpp>
-#include <orflib/products/bermudancallput.hpp>
 #include <orflib/methods/pde/pde1dsolver.hpp>
 
 using namespace std;
@@ -114,73 +113,6 @@ PY_BEGIN;
 
   // create the product
   orf::SPtrProduct spprod(new orf::AmericanCallPut(payoffType, strike, timeToExp));
-  // create the PDE solver
-  orf::Pde1DResults results;
-  bool storeAllResults = true;
-  orf::Pde1DSolver solver(spprod, spyc, spot, divYield, vol, results, storeAllResults);
-  solver.solve(pdeparams);
-
-  // write results
-  PyObject* ret = PyDict_New();
-  int ok = PyDict_SetItem(ret, asPyScalar("Price"), asPyScalar(results.prices[0]));
-
-  if (allresults) {
-    orf::Vector spots;
-    results.getSpotAxis(0, spots);
-    orf::Matrix values(results.times.size(), results.values.front().size());
-    for (size_t i = 0; i < results.times.size(); ++i)
-      for (size_t j = 0; j < results.values.front().size(); ++j)
-        values(i, j) = results.values[i](j, 0);
-
-    PyDict_SetItem(ret, asPyScalar("Times"), asNumpy(results.times));
-    PyDict_SetItem(ret, asPyScalar("Spots"), asNumpy(spots));
-    PyDict_SetItem(ret, asPyScalar("Values"), asNumpy(values));
-  }
-  return ret;
-
-PY_END;
-}
-
-
-
-static
-PyObject*  pyOrfBermBSPDE(PyObject* pyDummy, PyObject* pyArgs)
-{
-PY_BEGIN;
-
-  PyObject* pyPayoffType(NULL);
-  PyObject* pySpot(NULL);
-  PyObject* pyStrike(NULL);
-  PyObject* pyTimesToExer(NULL);
-  PyObject* pyDiscountCrv(NULL);
-  PyObject* pyDivYield(NULL);
-  PyObject* pyVolatility(NULL);
-  PyObject* pyPdeParams(NULL);
-  PyObject* pyAllResults(NULL);
-
-  if (!PyArg_ParseTuple(pyArgs, "OOOOOOOOO", &pyPayoffType, &pyStrike, &pyTimesToExer,
-    &pySpot, &pyDiscountCrv, &pyDivYield, &pyVolatility, &pyPdeParams, &pyAllResults))
-    return NULL;
-
-  int payoffType = asInt(pyPayoffType);
-  double spot = asDouble(pySpot);
-  double strike = asDouble(pyStrike);
-  orf::Vector timesToExer = asVector(pyTimesToExer);
-
-  std::string name = asString(pyDiscountCrv);
-  orf::SPtrYieldCurve spyc = orf::market().yieldCurves().get(name);
-  ORF_ASSERT(spyc, "error: yield curve " + name + " not found");
-
-  double divYield = asDouble(pyDivYield);
-  double vol = asDouble(pyVolatility);
-
-  // read the PDE parameters
-  orf::PdeParams pdeparams = asPdeParams(pyPdeParams);
-  // read the allresults flag
-  bool allresults = asBool(pyAllResults);
-
-  // create the product
-  orf::SPtrProduct spprod(new orf::BermudanCallPut(payoffType, strike, timesToExer));
   // create the PDE solver
   orf::Pde1DResults results;
   bool storeAllResults = true;